feat: 修复断点上传和更新上传路径

+++ utils/breakpoint_continue.go
@@ -1,6 +1,7 @@
 package utils
 
 import (
+	"errors"
 	ioutil "io"
 	"math"
 	"os"
@@ -123,6 +124,9 @@ func MakeFile(fileName string, FileMd5 string) (string, error) {
 //@return: error
 
 func RemoveChunk(FileMd5 string) error {
+	if FileMd5 == "" {
+		return errors.New("路径错误")
+	}
 	err := os.RemoveAll(breakpointDir + FileMd5) // RemoveAll删除path指定的文件，或目录及它包含的任何下级对象。它会尝试删除所有东西，除非遇到错误并返回。如果path指定的对象不存在，RemoveAll会返回nil而不返回错误。
 	return err
 }
